docs: create docs page
